refactored utils out of the project
