+++ lib/msal-core/src/AccessTokenKey.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { Utils } from "./utils/Utils";
+import { CryptoUtils } from "./utils/CryptoUtils";
 import { UrlUtils } from "./utils/UrlUtils";
 
 /**
@@ -18,6 +18,6 @@ export class AccessTokenKey {
     this.authority = UrlUtils.CanonicalizeUri(authority);
     this.clientId = clientId;
     this.scopes = scopes;
-    this.homeAccountIdentifier = Utils.base64Encode(uid) + "." + Utils.base64Encode(utid);
+    this.homeAccountIdentifier = CryptoUtils.base64Encode(uid) + "." + CryptoUtils.base64Encode(utid);
   }
 }
